smth wrong with JSON serialization

+++ Server/Services/ProductService.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Win32;
 using Server.Data;
 using Server.Services.ServiceInterfaces;
 using Shared;
@@ -13,29 +14,31 @@ namespace Server.Services
         {
             dbContextFactory = dbContext;
         }
-        public ProductDTO[] GetProducts()
+        public List<ProductDTO> GetProducts()
         {
             using var dbContext = dbContextFactory.CreateDbContext();
-            if(!dbContext.Products.Any()) 
-            {
-                return null;
-            }
+            //if(!dbContext.Products.Any()) 
+            //{
+            //    Console.WriteLine($"Joker");
+            //    return null;
+            //}
             List<ProductDTO> productsList = new List<ProductDTO>();
             foreach(var product in dbContext.Products)
             {
                 ProductDTO productDTO = new ProductDTO(product);
                 productsList.Add(productDTO);
+                Console.WriteLine($"Sended order #{productDTO.ProductName}");
             }
 
-            return productsList.ToArray();
+            return productsList;
         }
 
-        public ProductDTO AddProduct(ProductCreateDTO productCreateDTO)
+        public bool AddProduct(ProductCreateDTO productCreateDTO)
         {
             using var dbContext = dbContextFactory.CreateDbContext();
             if (dbContext.Products.Any(p => p.ProductName == productCreateDTO.ProductName))
             {
-                return null;
+                return false;
             }
             Product newProduct = new Product();
             newProduct.ProductName = productCreateDTO.ProductName;
@@ -43,29 +46,34 @@ namespace Server.Services
             dbContext.Products.Add(newProduct);
             dbContext.SaveChanges();
             ProductDTO responseProduct = new ProductDTO(newProduct);
-            return responseProduct;
+            return true;
         }
 
-        public ProductDTO EditProduct(long id, ProductDTO productDTO)
+        public bool EditProduct(long id, ProductDTO productDTO)
         {
             using var dbContext = dbContextFactory.CreateDbContext();
             if (!dbContext.Products.Any(p => p.Id == id))
             {
-                return null;
+                return false;
             }
             Product editableProduct = dbContext.Products.First(p => p.Id == id);
             editableProduct.ProductName = productDTO.ProductName;
             dbContext.SaveChanges();
             ProductDTO responseProduct = new ProductDTO(dbContext.Products.First(p => p.Id == id));
-            return responseProduct;
+            return true;
         }
 
-        public void DeleteProduct(long id)
+        public bool DeleteProduct(long id)
         {
             using var dbContext = dbContextFactory.CreateDbContext();
+            if (!dbContext.Products.Any(p => p.Id == id))
+            {
+                return false;
+            }
             Product deletableProduct = dbContext.Products.First(p => p.Id == id);
             dbContext.Products.Remove(deletableProduct);
             dbContext.SaveChanges();
+            return true;
         }
